refactor: use explicit runtime modules
