--- src/bridge/planner/__init__.py
@@ -1,5 +1,0 @@
-"""Deterministic construction of local BRIDGE analysis plans."""
-
-from bridge.planner.service import PlanBuilder
-
-__all__ = ["PlanBuilder"]
