--- src/bridge/storage/__init__.py
@@ -1,5 +1,0 @@
-"""Local persistence adapters for BRIDGE application artifacts."""
-
-from bridge.storage.artifacts import ArtifactVerification, LocalArtifactStore, StoredArtifact
-
-__all__ = ["ArtifactVerification", "LocalArtifactStore", "StoredArtifact"]
